qa-metrics-dashboard/backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
SQLAlchemy==2.0.36
psycopg2-binary==2.9.9
alembic==1.13.2
pydantic==2.9.2
python-multipart==0.0.9
passlib[bcrypt]==1.7.4
PyJWT==2.9.0
python-dotenv==1.0.1
scikit-learn==1.5.2
numpy==2.1.1
pandas==2.2.2
requests==2.32.3
pytest==8.3.3
